feat(layout): Layout is now in TS :)
